// server/src/types/chatTypes.ts
export type UUID = string;
// ...
export interface UserEntity {
  userId: UUID;
  createdAt: Date;
  profile: UserProfile;
  settings?: UserSettings;
}

export interface UserProfile {
  displayName: string;
  username: string;
  avatarUrl?: string;
  onlineStatus?: PresenceStatus;
  lastSeen?: Date;
  bio?: string;
}

export type PresenceStatus = 'Online' | 'Offline' | 'Away' | 'Busy' | 'Invisible' | 'Typing';

export interface UserSettings {
  notificationsEnabled?: boolean;
  readReceiptsEnabled?: boolean;
}